+++ src/lib/common/string.h
@@ -205,9 +205,8 @@ template <> std::string toString(float f);
 /*****************************************************************************
 *
 * isodate2str -
-*
 */
-extern std::string isodate2str(long long timestamp);
+extern std::string isodate2str(double timestamp);
